basic blockchain
- split modules into different files
- import relevant depedencies

[status: done]
